Start to share some common code between JavaScript files...
Do it little by little. A first attempt to share everything resulted in failure.

What is more, I saw a strange occurrence of parameters being passed as a query
string instead of in a POST... What on earth is going on here? Saw that in
heroku logs and locally, but don't know how to reproduce...